Rename and move test variables
* Rename the test variables to match their operation type.
* Rename files to match the operation type.
